Refactor Synchronizer
Move RepairContent and DeleteContent to other files
