+++ frugalos_segment/src/lib.rs
@@ -40,9 +40,11 @@ pub use service::{Service, ServiceHandle};
 pub mod config;
 
 mod client;
+mod delete;
 mod error;
 mod full_sync;
 mod metrics;
+mod repair;
 mod rpc_server;
 mod service;
 mod synchronizer;
