+++ frugalos_segment/src/synchronizer.rs
@@ -1,12 +1,10 @@
-use cannyls;
-use cannyls::deadline::Deadline;
 use cannyls::device::DeviceHandle;
-use cannyls::lump::LumpHeader;
 use fibers::time::timer::{self, Timeout};
 use frugalos_mds::Event;
 use frugalos_raft::NodeId;
 use futures::{Async, Future, Poll};
 use libfrugalos::entity::object::ObjectVersion;
+use libfrugalos::repair::RepairIdleness;
 use prometrics::metrics::{Counter, Gauge, Histogram, MetricBuilder};
 use slog::Logger;
 use std::cmp::{self, Reverse};
@@ -14,25 +12,22 @@ use std::collections::{BTreeSet, BinaryHeap};
 use std::env;
 use std::time::{Duration, Instant, SystemTime};
 
-use client::storage::{GetFragment, MaybeFragment, StorageClient};
-use config;
+use client::storage::StorageClient;
+use delete::DeleteContent;
 use full_sync::FullSync;
-use libfrugalos::repair::RepairIdleness;
+use repair::RepairContent;
 use service::{RepairLock, ServiceHandle};
-use util::Phase3;
 use Error;
 
 const MAX_TIMEOUT_SECONDS: u64 = 60;
 const DELETE_CONCURRENCY: usize = 16;
 
-type BoxFuture<T> = Box<dyn Future<Item = T, Error = Error> + Send + 'static>;
-
 // TODO: 起動直後の確認は`device.list()`の結果を使った方が効率的
 pub struct Synchronizer {
-    logger: Logger,
-    node_id: NodeId,
-    device: DeviceHandle,
-    client: StorageClient,
+    pub(crate) logger: Logger,
+    pub(crate) node_id: NodeId,
+    pub(crate) device: DeviceHandle,
+    pub(crate) client: StorageClient,
     service_handle: ServiceHandle,
     task: Task,
     // TODO: define specific types for two kinds of items and specialize the procedure for each todo queue
@@ -43,10 +38,10 @@ pub struct Synchronizer {
     enqueued_delete: Counter,
     dequeued_repair: Counter,
     dequeued_delete: Counter,
-    repairs_success_total: Counter,
-    repairs_failure_total: Counter,
-    repairs_unnecessary_total: Counter,
-    repairs_durations_seconds: Histogram,
+    pub(crate) repairs_success_total: Counter,
+    pub(crate) repairs_failure_total: Counter,
+    pub(crate) repairs_unnecessary_total: Counter,
+    pub(crate) repairs_durations_seconds: Histogram,
     full_sync_count: Counter,
     full_sync_deleted_objects: Counter,
     // How many objects have to be swept before full_sync is completed (including non-existent ones)
@@ -423,190 +418,5 @@ impl Future for Task {
             Task::Delete(ref mut f) => track!(f.poll()),
             Task::Repair(ref mut f, _) => track!(f.poll()),
         }
-    }
-}
-
-// #[derive(Debug)]
-struct DeleteContent {
-    futures: Vec<BoxFuture<bool>>,
-}
-impl DeleteContent {
-    pub fn new(synchronizer: &Synchronizer, versions: Vec<ObjectVersion>) -> Self {
-        debug!(
-            synchronizer.logger,
-            "Starts deleting contents: versions={:?}", versions
-        );
-
-        let futures = versions
-            .into_iter()
-            .map(move |v| {
-                let lump_id = config::make_lump_id(&synchronizer.node_id, v);
-                let future = synchronizer
-                    .device
-                    .request()
-                    .deadline(Deadline::Infinity)
-                    .delete(lump_id);
-                into_box_future(future)
-            })
-            .collect();
-        DeleteContent { futures }
-    }
-}
-impl Future for DeleteContent {
-    type Item = ();
-    type Error = Error;
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        let mut i = 0;
-        while i < self.futures.len() {
-            // NOTE: 一つ失敗しても全てを失敗扱いにする必要はない
-            if let Async::Ready(_) = track!(self.futures[i].poll().map_err(Error::from))? {
-                self.futures.swap_remove(i);
-            } else {
-                i += 1;
-            }
-        }
-        if self.futures.is_empty() {
-            Ok(Async::Ready(()))
-        } else {
-            Ok(Async::NotReady)
-        }
-    }
-}
-
-fn into_box_future<F>(future: F) -> BoxFuture<F::Item>
-where
-    F: Future<Error = cannyls::Error> + Send + 'static,
-{
-    Box::new(future.map_err(Error::from))
-}
-
-// NOTE
-// ====
-//
-// 以下の処理を行う:
-// 1. `version`に対応するオブジェクトの中身が存在するかチェック
-// 2. 存在しないなら復元を試みる (以下は一番複雑なdispersedの場合のみの処理を記述)
-// 2-1. 一つ前のクラスタ構成での対応するノードからセグメントを移動する (クラスタ構成変更時; 未実装)
-//      => 移動が完了したら、古いノードのセグメントは削除する
-// 2-2. それが無理なら、クラスタ内の他のノードからセグメント群を集めて復元する
-//
-// なおバケツの種別が`metadata`の場合には、復元は不要.
-// `replicated`の場合には、クラスタ内の任意の一つのノードからコピーすれば良い
-// (クラスタ構成変更によって、完全にノード構成が変わった場合にだけ注意が必要).
-struct RepairContent {
-    logger: Logger,
-    node_id: NodeId,
-    version: ObjectVersion,
-    client: StorageClient,
-    device: DeviceHandle,
-    started_at: Instant,
-    repairs_success_total: Counter,
-    repairs_failure_total: Counter,
-    repairs_unnecessary_total: Counter,
-    repairs_durations_seconds: Histogram,
-    phase: Phase3<BoxFuture<Option<LumpHeader>>, GetFragment, BoxFuture<bool>>,
-}
-impl RepairContent {
-    pub fn new(synchronizer: &Synchronizer, version: ObjectVersion) -> Self {
-        let logger = synchronizer.logger.clone();
-        let device = synchronizer.device.clone();
-        let node_id = synchronizer.node_id;
-        let lump_id = config::make_lump_id(&node_id, version);
-        let started_at = Instant::now();
-        let repairs_success_total = synchronizer.repairs_success_total.clone();
-        let repairs_failure_total = synchronizer.repairs_failure_total.clone();
-        let repairs_unnecessary_total = synchronizer.repairs_unnecessary_total.clone();
-        let repairs_durations_seconds = synchronizer.repairs_durations_seconds.clone();
-        debug!(
-            logger,
-            "Starts checking content: version={:?}, lump_id={:?}", version, lump_id
-        );
-        let phase = Phase3::A(into_box_future(
-            device.request().deadline(Deadline::Infinity).head(lump_id),
-        ));
-        RepairContent {
-            logger,
-            node_id,
-            version,
-            client: synchronizer.client.clone(),
-            device,
-            started_at,
-            repairs_success_total,
-            repairs_failure_total,
-            repairs_unnecessary_total,
-            repairs_durations_seconds,
-            phase,
-        }
-    }
-}
-impl Future for RepairContent {
-    type Item = ();
-    type Error = Error;
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        while let Async::Ready(phase) = track!(self.phase.poll().map_err(|e| {
-            self.repairs_failure_total.increment();
-            e
-        }))? {
-            let next = match phase {
-                Phase3::A(Some(_)) => {
-                    debug!(self.logger, "The object {:?} already exists", self.version);
-                    self.repairs_unnecessary_total.increment();
-                    return Ok(Async::Ready(()));
-                }
-                Phase3::A(None) => {
-                    debug!(
-                        self.logger,
-                        "The object {:?} does not exist (try repairing)", self.version
-                    );
-
-                    let future = self.client.clone().get_fragment(self.node_id, self.version);
-                    Phase3::B(future)
-                }
-                Phase3::B(MaybeFragment::NotParticipant) => {
-                    debug!(
-                        self.logger,
-                        "The object {:?} should not be stored on this node: node_id={:?}",
-                        self.version,
-                        self.node_id
-                    );
-                    self.repairs_failure_total.increment();
-                    return Ok(Async::Ready(()));
-                }
-                Phase3::B(MaybeFragment::Fragment(mut content)) => {
-                    ::client::storage::append_checksum(&mut content); // TODO
-
-                    let lump_id = config::make_lump_id(&self.node_id, self.version);
-                    debug!(
-                        self.logger,
-                        "Puts repaired content: version={:?}, lump_id={:?}, content_size={}",
-                        self.version,
-                        lump_id,
-                        content.len()
-                    );
-
-                    let data = track!(self.device.allocate_lump_data_with_bytes(&content))
-                        .expect("TODO: error handling");
-                    let future = self
-                        .device
-                        .request()
-                        .deadline(Deadline::Infinity)
-                        .put(lump_id, data);
-                    Phase3::C(into_box_future(future))
-                }
-                Phase3::C(_) => {
-                    debug!(
-                        self.logger,
-                        "Completed repairing content: {:?}", self.version
-                    );
-                    self.repairs_success_total.increment();
-                    let elapsed =
-                        prometrics::timestamp::duration_to_seconds(self.started_at.elapsed());
-                    self.repairs_durations_seconds.observe(elapsed);
-                    return Ok(Async::Ready(()));
-                }
-            };
-            self.phase = next;
-        }
-        Ok(Async::NotReady)
     }
 }
